Risolti alcuni bug ma non tutti

+++ src/main/java/it/contrader/dao/SchedaVotazioneRepository.java
@@ -7,6 +7,7 @@ import it.contrader.model.SchedaVotazione;
 
 @Repository
 @Transactional
-public interface SchedaVotazioneRepository extends CrudRepository<SchedaVotazione, Integer> {
+public interface SchedaVotazioneRepository extends CrudRepository<SchedaVotazione, Long> {
+	
 	
 }
